Absolute values are not complex, la!

+++ libsquiggly/analysis/rolling_abs_mean.py
@@ -23,7 +23,7 @@ def rolling_abs_mean(data, N, mode="same"):
 		The mean of the window at its current shift in the data stream
 	"""
 
-	rb = RollingBuffer(N, dtype=complex64)
+	rb = RollingBuffer(N, dtype=float64)
 	# Pre-load so that we're operating in "valid" mode
 	for idx in xrange(N-1):
 		rb.push(abs(next(data)))
